third_party/WebKit/LayoutTests/http/tests/devtools/elements/edit/edit-dom-actions-shadow-2-expected.txt

Tests that user can mutate author shadow DOM by means of elements panel.


Running: testSetUp

Running: testEditShadowDOMAsHTML
==== before ====
- <div id="testEditAuthorShadowDOMAsHTML">
    - #shadow-root (open)
          <div id="authorShadowDOMElement"></div>
  </div>
<div id="authorShadowDOMElement"></div>
==== after ====
- <div id="testEditAuthorShadowDOMAsHTML">
    - #shadow-root (open)
        - <span foo="shadow-span">
              <span id="inner-shadow-span">Shadow span contents</span>
          </span>
  </div>
